+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 2.54.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 2.53.0 - (March 25, 2024)
 
 * Changed